Fixed issue in race simulation where car wouldn't stop at control stops.

+++ RaceSim/include/route/route.h
@@ -23,7 +23,7 @@ protected:
     std::vector<double> segment_lengths;
 
     /* Indices of all control stops */
-    std::unordered_set<uint32_t> control_stops;
+    std::unordered_set<size_t> control_stops;
 
     /* Points of the route */
     std::vector<Coord> route_points;
@@ -45,7 +45,7 @@ public:
     uint32_t get_num_segments() const;
     std::vector<std::pair<uint32_t, uint32_t>> get_segments() const;
     std::vector<double> get_segment_lengths() const;
-    std::unordered_set<uint32_t> get_control_stops() const;
+    std::unordered_set<size_t> get_control_stops() const;
     std::vector<Coord> get_route_points() const;
     uint32_t get_num_points() const;
 
@@ -53,7 +53,7 @@ public:
     void set_num_segments(uint32_t num);
     void set_segments(std::vector<std::pair<uint32_t, uint32_t>> seg);
     void set_segment_lengths(std::vector<double> seg_lengths);
-    void set_control_stops(std::unordered_set<uint32_t> c_stops);
+    void set_control_stops(std::unordered_set<size_t> c_stops);
     void set_route_points(std::vector<Coord> new_route_points);
 };
